Create tic tac toe app without computer logic
The app currently has two scenes. The first scene allows the user to choose "x" or "o" as her icon. The second scene draws the tic tac toe gameboard to the canvas. After the gameboard is drawn to the canvas the user can draw their icon on the gameboard. The computer's logic is not set up.